pyproject.toml
[project]
name = "qa-agent"
version = "0.1.0"
description = "QA Agent — automate post-regression triage for DV engineers."
authors = [{name="Pravab Kar"}]
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk",
    "anthropic>=0.30.0",
    "openai",
    "google-genai",
    "python-dotenv",
    "platformdirs",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
qa-agent = "qa_agent.cli:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
